Move Kolibri environment setup to another module
https://phabricator.endlessm.com/T34007
