+++ src/kolibri_android/kolibri_utils.py
@@ -1,10 +1,7 @@
 import logging
 import os
 import re
-import sys
-from pathlib import Path
 
-from .android_utils import apply_android_workarounds
 from .android_utils import get_android_node_id
 from .android_utils import get_endless_key_uris
 from .android_utils import get_home_folder
@@ -13,22 +10,10 @@ from .android_utils import get_timezone_name
 from .android_utils import get_version_name
 from .android_whitenoise import DynamicWhiteNoise  # noqa: E402
 from .android_whitenoise import monkeypatch_whitenoise  # noqa: E402
+from .globals import SCRIPT_PATH
 
-SCRIPT_PATH = Path(__file__).absolute().parent.parent
 
-
-def initialize():
-    # initialize logging before loading any third-party modules, as they may cause logging to get configured.
-    logging.basicConfig(level=logging.DEBUG)
-    jnius_logger = logging.getLogger("jnius")
-    jnius_logger.setLevel(logging.INFO)
-
-    apply_android_workarounds()
-
-    sys.path.append(SCRIPT_PATH.as_posix())
-    sys.path.append(SCRIPT_PATH.joinpath("kolibri", "dist").as_posix())
-    sys.path.append(SCRIPT_PATH.joinpath("extra-packages").as_posix())
-
+def init_kolibri():
     monkeypatch_whitenoise()
 
     signing_org = get_signature_key_issuing_organization()
