+++ src/workers.py
@@ -1,5 +1,7 @@
-from kolibri_android.initialization import initialize
+from kolibri_android.globals import initialize
+from kolibri_android.kolibri_utils import init_kolibri
 
 initialize()
+init_kolibri()
 
 from kolibri_android.workers_service import *  # noqa F403 F401
